requirements.txt
absl-py==0.10.0
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.0.4
astunparse==1.6.3
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
cachetools==4.1.1
certifi==2018.8.24
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
defusedxml==0.5.0
distlib==0.3.1
dj-database-url==0.5.0
Django==2.2.10
django-allauth==0.37.1
django-cors-headers==2.4.0
django-rest-auth==0.9.3
django-tinymce4-lite==1.8.0
djangorestframework==3.9.1
entrypoints==0.3
filelock==3.0.12
gast==0.3.3
google-auth==1.21.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==19.7.1
h5py==2.10.0
idna==2.7
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.4
jedi==0.17.2
Jinja2==2.11.2
jsmin==2.2.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.3.1
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.0
nbconvert==6.0.5
nbformat==5.0.7
nest-asyncio==1.4.0
notebook==6.1.4
numpy==1.18.5
oauthlib==2.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
pipenv==2020.8.13
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.1
pylint==2.1.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2018.5
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.21.0
requests-oauthlib==1.0.0
rsa==4.6
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.3.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tornado==6.0.4
traitlets==5.0.4
typed-ast==1.1.0
urllib3==1.24.2
virtualenv==20.0.29
virtualenv-clone==0.5.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==3.3.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
